+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 -----------------------
 
 * Fix a bug in :meth:`pymaf.MafFrame.plot_mutated` when using the ``hue`` option.
+* Add ``sort`` argument to :meth:`pymaf.MafFrame.plot_vaf` method.
 
 0.18.0 (2021-07-20)
 -------------------